Extract grayscale imagesource
